Mod how Players are represented. Impl MOVE Enum
Player is tracked in a variable. Removed the representation of the player from the world state.
GridWorld is now a static allocation of free, tiles, start, and goal.
Player position is visualised on print over-top of existing map rather than interfering.
This prevents the need for a 'list' per Tile approach, which would be needed if Player 'lived' in the world.

All checks for player utilise player variable for tracking, rather than specific Enum at Tile location.